update tests to point to the right place post psf moves
