+++ photutils/psf/tests/test_misc.py
@@ -7,7 +7,7 @@ from __future__ import division
 import numpy as np
 from numpy.testing import assert_allclose
 from astropy.tests.helper import pytest
-from ..core import IntegratedGaussianPRF, prepare_psf_model
+from .. import IntegratedGaussianPRF, prepare_psf_model
 
 try:
     HAS_SCIPY = True
